% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwann.R
\name{nnet}
\alias{nnet}
\title{Build a basic artifical neural network (experimental).}
\usage{
nnet(
  x_train,
  y_train,
  x_pred,
  norm = T,
  nrHidden = ncol(x_train) * 2,
  batchSize = 50,
  optimizer = "nesterov",
  lr = 0.05,
  linOut = T,
  iterations = NA,
  cv_max_iterations = Inf,
  cv_patience = 999,
  cv_folds = 10,
  cv_repeats = 1,
  permutations = 0,
  threads = 4
)
}
\description{
Build a basic artifical neural network (experimental).
}
