% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwann.R
\name{predict}
\alias{predict}
\title{Predict values for a trained GWANN model.}
\usage{
predict(gwann_o, x_pred)
}
\arguments{
\item{gwann_o}{GWANN model as Java object.}

\item{x_pred}{Matrix of prediction data. Rows are observations, columns are independent variables.}
}
\description{
Locations of observations must match with locations of the prediction data when the model was built.
}
