% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwann.R
\name{predict_gwann}
\alias{predict_gwann}
\title{Predict values for a trained GWANN/NNet model (experimental).}
\usage{
predict_gwann(model, x_pred)
}
\arguments{
\item{x_pred}{Matrix of prediction data. Rows are observations, columns are independent variables.}

\item{ro}{GWANN model as Java object.}
}
\description{
Locations of observations must match with locations of the prediction data when the model was built.
}
