% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwann.R
\name{gwann}
\alias{gwann}
\title{Build a Geographically Weighted Artificial Neural Network.}
\usage{
gwann(
  x,
  y,
  dm,
  trainIdx = 1:nrow(dm),
  predIdx = 1:nrow(dm),
  nrHidden = 4,
  batchSize = 10,
  optimizer = "nesterov",
  lr = 0.1,
  linOut = T,
  kernel = "gaussian",
  bandwidth = NA,
  adaptive = F,
  gridSearch = F,
  minBw = NA,
  maxBw = NA,
  steps = 20,
  iterations = NA,
  patience = 100,
  folds = 10,
  repeats = 1,
  threads = 4
)
}
\arguments{
\item{x}{Matrix. Rows are observations, columns are independent variables.}

\item{y}{Vector. Values represent target values for the observations in \code{x}.}

\item{dm}{Matrix of distances between the observations of \code{x}.}

\item{trainIdx}{a vector containing the indices of observations of \code{x} used for training.}

\item{predIdx}{a vector containing the indices of observations of \code{x} used for prediction.}

\item{nrHidden}{Number of hidden neurons.}

\item{batchSize}{Batch size.}

\item{optimizer}{Optimizer (sgd, momentum, nesterov).}

\item{lr}{Learning rate.}

\item{kernel}{Kernel.}

\item{bandwidth}{Bandwidth size. If NA, it is determined using 10-fold CV.}

\item{adaptive}{Adaptive instead of fixed bandwidth?}

\item{gridSearch}{If bandwidth is NA, use grid search or local optimization for finding appropriate bandwidth.}

\item{minBw}{Lower limit for bandwidth search.}

\item{maxBw}{Upper limit for bandwidth search.}

\item{steps}{Number of bandwidths to test when doing a grid search.}

\item{iterations}{Number of training iterations. If NA, is determined using 10-fold CV.}

\item{patience}{After how many iterations with no improvement should training prematurely stop?}

\item{folds}{Number of cross-validation folds}

\item{repeats}{Number of repeats of cross-validation procedure}

\item{threads}{Number of threads to use.}
}
\value{
A list of five elements.
The first element \code{predictions} contains the predictions.
The second elemnt \code{weights} contains the connection weights of the hidden neurons to the output neurons.
The third element \code{rmse} is the mean RMSE of the CV procedure.
The fourth element \code{bandwidth} is the bandwidth that is used to train the final model.
The fifth element \code{iterations} is the numer of training iterations for the final model.
}
\description{
Build a Geographically Weighted Artificial Neural Network.
}
\examples{
data(toy4)

dm<-as.matrix( dist(toy4[,c("lon","lat")])  )
x<-as.matrix(toy4[,c("x1","x2")])
y<-as.numeric( toy4[,c("y")] )

\dontrun{
r<-gwann(x=x,y=y,dm=dm,trainIdx=1:nrow(x),predIdx=1:nrow(x),nrHidden=5,batchSize=50,bandwidth=1.8,iterations=5610,lr=0.01)

if( all ( sapply( c("ggplot2","reshape2","viridis"), require, character.only=T ) ) ) {
   s<-cbind( Prediction=diag(r$predictions), toy4[,c("lon","lat")] )
   ggplot(s,aes(lon,lat,fill=Prediction)) + geom_raster() + scale_fill_viridis() + coord_fixed()
}
}
}
\references{
Not yet published
}
